feature/invite-users: separate methods from business logic
