+++ imports/modules/organizations/client/service/organization.service.js
@@ -1,14 +1,16 @@
+import { ORGANIZATION_METHODS } from '../../server/methods/origanization.server.methods';
+
 class OrganizationServiceClass {
 	constructor() {
 	}
 
 	async sendCreatingOrganizationRequest({ title }) {
-		return await Meteor.callPromise('organization.methods.createOrganizationRequest', { title })
+		return await Meteor.callPromise(ORGANIZATION_METHODS.createOrganizationRequest, { title })
 	}
 
 
 	async approveOrganization(organizationId, ownerId, groupTitle) {
-		return await Meteor.callPromise('organization.methods.approveOrganization', {
+		return await Meteor.callPromise(ORGANIZATION_METHODS.approveOrganization, {
 			organizationId,
 			ownerId,
 			groupTitle
